+++ src/main/java/com/teamharmony/newscommunity/auth/config/security/AuthConstants.java
@@ -1,4 +1,4 @@
-package com.teamharmony.newscommunity.security;
+package com.teamharmony.newscommunity.auth.config.security;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
