+++ src/main/java/com/teamharmony/newscommunity/auth/filter/CustomAuthenticationFilter.java
@@ -1,10 +1,10 @@
-package com.teamharmony.newscommunity.filter;
+package com.teamharmony.newscommunity.auth.filter;
 
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.teamharmony.newscommunity.users.entity.Tokens;
-import com.teamharmony.newscommunity.users.repository.TokensRepository;
+import com.teamharmony.newscommunity.auth.entity.Tokens;
+import com.teamharmony.newscommunity.auth.repository.TokensRepository;
 import org.springframework.http.ResponseCookie;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
