+++ src/main/java/com/teamharmony/newscommunity/auth/filter/CustomAuthorizationFilter.java
@@ -1,12 +1,12 @@
-package com.teamharmony.newscommunity.filter;
+package com.teamharmony.newscommunity.auth.filter;
 
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.JWTVerifier;
 import com.auth0.jwt.algorithms.Algorithm;
 import com.auth0.jwt.interfaces.DecodedJWT;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.teamharmony.newscommunity.security.AuthConstants;
-import com.teamharmony.newscommunity.users.repository.TokensRepository;
+import com.teamharmony.newscommunity.auth.config.security.AuthConstants;
+import com.teamharmony.newscommunity.auth.repository.TokensRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
