+++ src/main/java/com/teamharmony/newscommunity/auth/repository/TokensRepository.java
@@ -1,7 +1,7 @@
-package com.teamharmony.newscommunity.users.repository;
+package com.teamharmony.newscommunity.auth.repository;
 
 
-import com.teamharmony.newscommunity.users.entity.Tokens;
+import com.teamharmony.newscommunity.auth.entity.Tokens;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TokensRepository extends JpaRepository<Tokens, Long> {
